build: linux: Place generated files in OUT_DIR
When cross-compiling the eventmsgs.lib in source control cannot be used
as it's incompatible with the mingw toolchain. Therefore it is freshly
generated in the source directory on every build since
  cc0be9c Always run windmc/windrc on MinGW.

This becomes a problem when using the eventlog crate in project with
locally vendored dependencies. Since changes the situation by placing
the build artifacts in OUT_DIR.

Signed-off-by: Konrad Gräfe <kgraefe@paktolos.net>